Revert "Revert "Prefer double-quoted strings in Ruby code""
This reverts commit 0bb7ae20d70ddc0b21629d34f45d2fa5a8054ea7.

Commit in mediawiki/selenium repository can not break Jenkins builds. Jenkins builds see this code only _after_ a new release of mediawiki-selenium gem is made _and_ a repository updates the version of the gem it uses.

We have found and fixed[1] the problem that caused all Jenkins jobs to break.

Change-Id: I70f0f6dc92b8f77f472fad0446ea04c3ea2f5481
1: http://lists.wikimedia.org/pipermail/qa/2013-December/000802.html