Support for default environment configurations
Default configuration for different types of environments can now be
defined in an `environments.yml` file. The contents of the file should
be a YAML hash containing key/hash pairs corresponding to a environment
name and set of environment variable defaults. If a value for
`MEDIAWIKI_ENVIRONMENT` is provided, the set of defaults defined in the
file will be merged with any given via environment variables directly,
the latter taking precedence.

Change-Id: Ifb358ec205b36ff0fbde9b4b16501f5ed354c28c