Refactored EAL configuration overrides
One big caveat of implementing `Environment#with` by cloning + using
instance evaluation over block evaluation with a normal closure is that
instance variables set within the scope of the given block are lost to
the original `Environment` object.

  Step(...) do
    @foo = "a"
    with_alternative(...) do # or #on_wiki, #as_user, etc.
      @foo = "b"
    end
    @foo == "b" # => false
  end

This behavior would likely violate the principle of least surprise for
most users, so the cloning has been replaced by temporarily overwriting
environment config state, yielding normally (so the expected closure is
created), and restoring the original state.

Change-Id: I3444720eae3ff1d388e39529cb3589f786ead0ec