Remove explicit onfocus
This is basically the root cause of T87446.

This was put in originally to prevent mis-clicks on the "Log in" button
from early days of the new UI redesign. Let's hope the parts of the
element are less susceptible to that today.

I have been told that the latest Firefox will no longer support
native events without e.g. building specially from source.

So let's remove 'onfocus', update the repos and see how it goes.

Change-Id: Id6be3687e49b715ef038fbe0fa72a4ffe463a6b7