User factory for automatic creation of test accounts
Implemented `UserFactory` and `UserFactoryHelper` to aide with the
creation of randomized fixture accounts in environments where a greater
level of atomicity is desired and the creation of many throwaway
accounts is not an issue (namely, in CI).

The feature can be enabled per environment by setting `user_factory` to
`true` or `false` in `environments.yml`.

Refactored `Environment#as_user` to delegate username/password lookup to
`Environment#user` and `Environment#password`, methods that are
reimplemented by the factory helper.

Bug: T103135
Change-Id: Ia4dbfc3a225b9a514d65ded16075ec302140adc2