Video recording of headless execution
Moved headless implementation to a new `HeadlessHelper` module so it's
independent of Cucumber and can be tested more easily. The module
extends `Environment` to automatically create a new headless display the
first time a browser is ever requested.

Video recording is started automatically if `headless_capture_path` has
been defined, and saved under that path if the scenario fails. Video
files for passing scenarios are discarded.

Updated headless dependency to get support for libav. Recording uses
`avconv` under the hood and the libx264 codec with an mp4 container for
maximum compression and compatibility.

Paired with Zeljko.

Bug: T104583
Change-Id: I56f47884adde3b201a5cffc6e35915243db33661