Login helper for fast API-based authentication
The `LoginHelper` provides a `log_in` method that authenticates via the
API and adds the resulting cookies to the current browser instance.

This technique is a much faster way to satisfy an initial logged in
precondition and should speed up a large number of end-to-end MediaWiki
tests. It also has the added benefit of not exposing credentials in the
Selenium log.

Bug: T112908
Change-Id: I66002f5b824c2097effb8f9feda42e99b01074ba