Decouple screenshot-ing and artifacts from Cucumber hooks
Created `ScreenshotHelper` to encapsulate and generalize the
screenshot-ing previously implemented directly in a Cucumber `After`
hook.

Refactored `Environment#teardown` and friends to return artifacts in the
form `{ path => mime_type }` which can then be logged in the way
appropriate to the test runner.

Moved logging of SauceLabs job URL to a `RemoteBrowserFactory#teardown`
and refactored `HeadlessHelper` for the new artifacts system.

Bug: T108273
Change-Id: I4534fcb05180419ee9aa354f12616aadec160659