Never show firstrun pages in Firefox
I believe that no test should ever show any of Mozillas "firstrun"
pages. If such a page is shown and which is shown depends on the
current Firefox version and the time when and how often it was started
before with the same profile. These details should not be part of any
tests.

These "this is the first time you run a major update of Firefox" pages
do no actual harm, but it can be extremely confusing to see them,
especially when the first test fails and you think the reason is that
Firefox loaded the wrong page.

Change-Id: I7277ad771b329225df17673f101f73a14388b083