Provide Rake task to serve as a CI entrypoint
Implemented a Rake task that runs Cucumber in a standard way with
respect to our Jenkins configuration.

So one can invoke Rake task from anywhere in the working copy,
Environment learned 'test_dir', the base directory from which to lookup
for environments.yml and tests/features.  If relative, paths are made
absolute so the Environment properly load the YAML file and to have
cucumber to point to the proper place.

Add bunch of spec to cover the test_dir parameter.

Update README.md with a couple examples to add the rake task.

Bug: T128190
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: I627d0603487ab88e375fe5aa4fca2f8bb2a07790